tests: Move channel access tests to test_channel_access.
